move to antlr4.6

+++ src/syntaxErrorLogger.cpp
@@ -4,7 +4,7 @@ void SyntaxErrorLogger::syntaxError(
 		IRecognizer *recognizer,
 		Token *offendingSymbol,
 		size_t line,
-		int charPositionInLine,
+		size_t charPositionInLine,
 		const std::string &msg,
 		std::exception_ptr e) {
 	std::cerr << line << ":" << charPositionInLine << ":Error:" << msg << "\n";
@@ -26,7 +26,7 @@ void SyntaxErrorLogger::reportContextSensitivity(
 		const dfa::DFA &dfa,
 		size_t startIndex,
 		size_t stopIndex,
-		int prediction,
+		size_t prediction,
 		atn::ATNConfigSet *configs) {
 
 }
